+++ Tp_final - Server/Tp_final/LogicFSM.h
@@ -138,7 +138,7 @@ public:
 	void reset_game();
 
 	//saved_EventPackages
-	EventPackage * give_me_the_saved_EventPackage(unsigned int ID);
+	EventPackage * give_me_the_saved_EventPackage(uint16_t ID);
 	void saved_an_EventPackage(EventPackage * package_to_be_saved);
 
 protected:
@@ -149,7 +149,7 @@ protected:
 	LogicEventGenerator * ev_gen = NULL;
 private:
 
-	map<unsigned int, EventPackage*> saved_EventPackages;						// map of saved eventpackages
+	map<uint16_t, EventPackage*> saved_EventPackages;						// map of saved eventpackages
 
 	EventPackageFactory ev_pack_factory;
 	PackageFactory pack_factory;
